Fix merge bustage.
UltraBlame original commit: 49dad96faabaa36d24e85eb379e70bc31dd14ac5

+++ js/src/ion/x64/CodeGenerator-x64.cpp
@@ -325,7 +325,7 @@ CodeGeneratorX64::visitRecompileCheck(LRecompileCheck *lir)
     
     
     
-    const size_t *useCount = gen->info().script()->addressOfUseCount();
+    const uint32_t *useCount = gen->info().script()->addressOfUseCount();
     masm.movq(ImmWord(useCount), ScratchReg);
 
     Operand addr(ScratchReg, 0);
